chore(zero-client): add alive clients manager (#4508)
A class that lists the alive clients in a client group. It uses the
`navigator.locks` API to manage locks for each client. The class is
designed to be used in a browser environment where the `navigator.locks`
API is available.

When `navigator.locks` is not available (node.js, deno, bun),
it will return a set only containing the current `clientID`.

It uses one lock per client, identified by a combination of
`clientGroupID` and `clientID`. Then the `query` method is used to get
the list of all clients that hold or are waiting for locks in the same
client group.