feat(zero-react): ZeroProvider can now construct Zero too.
ZeroProvider can now be passed props as well as a Zero instance.

If passed props, ZeroProvider creates a Zero instance in an effect
so that it only happens on the client. ZeroProvider also takes care
of closing the Zero instance at the right time and creating a new
instance when any props change.