feat(zero): Limit TTL to 10 minutes (#4500)
If the TTL is set to "never" or a value greater than 10 minutes, we now
default it to 10 minutes. This is to prevent that queries are active for
too long, which can lead to performance issues and resource exhaustion.

This is dealt with on both the client and server. On the client we log
a warning. On the server we treat `-1` as 10 minutes so that old
existing queries will also expire.