fix(zero): Inspector queries were broken (#4558)
When we added custom queries we changed the type of the ast column to
nullable but the inspector code was not updated to handle that. This
commit updates the inspector code to handle the nullable ast column as
well as the name and args columns that were added when custom queries
were added.