refactor(zero-solid): reconcile changes to data due to query changes (#4559)
Starting from this PR contribution: https://github.com/rocicorp/mono/pull/4474

Prior to this change the zero-solid integration created a new store whenever a query changed.
This caused everything depending on the store to be disposed and recreated, instead of receiving granular updates.
This PR ensures that every `createQuery` call only creates a store the first time and passes on the setter to the view, and the view is responsible for reconciling data when there's large changes.

- Add a `ZeroProvider` and `useZero` so that the zero instance can be accessed through context, `ZeroProvider` can take a signal for a `ZeroOptions` or a `Zero` instance.
- `useZero` return a signal of `Zero` instance since the signal to ZeroProvider can change the `ZeroOptions` or `Zero` instance.
- Un-deprecate `useQuery` since it now requires context. Deprecates `createQuery` instead. `useX` is a good convention to communicate that this should be called inside a component and not outside.
- Added `createUseZero` to prevent redundant annotations, and so the api mirror's React's apis much more closely.

Next steps:
- Use solid's `reconcile` utility to achieve granular updates for they initial hydration of a changed query.  This will require adding synthetic ids to rows.
- Update docs