fix(zql): fix case where an edit to the bound node caused us to go over the limit (#4563)
editing the bound node for `take` would not remove the bound node! This caused us to go over the limit.

Not only go over the limit but to go over the limit and keep stale data in the view as the bound node would also not be updated to the new value.