chore(zero): fix race with mutation resolution (#4608)
The big picture is that we should have gone with writing mutation responses to the database in the first place.

The specific problem --

A user's API server could return an error but have applied the mutations. In this case, the mutation tracker was assuming that the mutations were not applied and would wait for a retry to apply them.

This would never happen since the mutations were applied with the original request.

The fix is to mark error responses to pushes as being in an unknown or limbo state.
When the lmid is advanced, we resolve limbo mutations whose lmid is less than or equal to the current lmid.