chore(zero): workaround for custom upstream (#4684)
If the upstream is custom we cannot try to delete mutation results from it.

Temporary workaround to not break users with custom upstreams.

Alternative would be to send a message to the API server when a mutation is acked, client is deleted or client group is deleted. Another alternative is to add more to the custom upstream interface and poke that. A third route would be to have the API server clean the table on some interval given zero is not expected to run for prolonged periods offline.