poc: use map by PK to speed up PK lookups.
Increases perf of chinook pk lookup test by 7x (18µs to 2.7µs).

Another 2x increase (to 1.3µs) is possible if we are allowed to
return row by ref, but I am not sure if this is allowed.

Note we can use same trick in MemorySource as it is 1.7x slower
(due to b-tree lookups).