fix(analyze-query): use replica based tablespecs TableSource schemas to match zero-cache behavior (#4822)
Another step towards making analyze-query's behavior more perfectly match zero-cache.

Without this change some queries that would succeed against zero-cache would fail with an exception with analyze-query.

In particular, if 
1. zero-cache determine that a column C should be part of a table T's primary key because it has a unique index on it
2. column C does not appear in schema.ts for table T

This caused queries involving the 'user' table of zbugs to fail, because of the `githubID` column, which has a unique index on it, but is not in schema.ts.