feat(zero-client): Use admin password for inspector (#4825)
Add support for admin password authentication in the inspector.

The server can now respond to any inspect rpc with `{op: 'authenticated', value: false}`. If
the client receives this, it will prompt the user for the admin password
and send a `{op: 'authenticate', value: 'password'}` rpc request. When a successful auth response is
received, the client then retries the original inspect rpc.

On the server side, the `InspectorDelegate` keeps tracks of authenticated
client groups by keeping a set in memory.