fix(expo-sqlite): do not close a db we're trying to open (#4851)
The `SQLiteDatabaseManager` has a map of instances that it keys by database name. When someone tells the SQLiteDBManager to close a db by name, it looks up all connections that exist for that DB and closes them.

This was a problem when one Zero instance was being destroyed and another was being created. Both instances use the `replicache_dbs_v0`. What was happening is:

Zero Instance 1 would be closing.
Zero Instance 2 would be opening.

(note: zero.close is async but it is not / can not be awaited in useEffect in ZeroProvider)

The `replicache_dbs_v0` would be opened by Zero Instance 2.
The close from Zero Instance 1 would come in.
The SQLiteDBManager would close all connections associated with `replicache_dbs_v0` which included Zero Instance 2's connections.

We would then fail when making any call to SQLite on that connection.

```
🟠 FunctionCallException: Calling the 'runSync' function has failed (at ExpoModulesCore/SyncFunctionDefinition.swift:137)
🟠 → Caused by: AccessClosedResourceException: Access to closed resource (at ExpoSQLite/SQLiteModule.swift:625)
```

This error does not show in the Expo logs but does show when:

a. running the app in xcode or
b. opening `Console.app` and selecting your simulator

---

The fix is to not have a single `SQLiteDBManager` but to create one for each Store. When the store is closed it then only closes its own connections to SQLite rather than the connections of other stores.

https://rocicorp.slack.com/archives/C013XFG80JC/p1756914030938379