feat(zero): Add support for StandardSchema to synced queries.
This was surprisingly difficult because StandardSchema supports
async validation. This meant that SyncedQuery needed to get forked
into ValidatedQuery, but fortunately that didn't spiral too far.

The other crazy unfortunate collision was that the async
validation means that `withValidation` becomes async. But the return
value `withValidation` is a query. And queries themselves are
promise-like. This means that when code that calls `withValidation`
uses await, it will unwrap recursively because that's what await does.

The result is that awaiting the result of `withValidation` in order to
return the query over the http response actually tries to run the
query.

The one lucky break was that `handleGetQueriesRequest` actually wants
{query: Q}, not Q as an input. So I was able to leverage that and
have `withValidation` return that shape not a raw query. This stops
the recursive unwrapping of promises from happening. I think the
previous API was actually better, but it's the best I could think
of.

We probably need to remove the ability to run queries directly, as
we've now added zero.run and friends. But we cannot do that until we
add tx.run. That will likely have usability impacts on reading data
inside transactions.