Change DeletedClients to use an array ID "pairs" (#4882)
This refactors the handling of deleted clients on the client side. Instead of tracking deleted clients by separate IDs, we now use tuples of `{clientGroupID, clientID}`. This change allows for more precise filtering: only deleted clients from the current client group are sent to the server.

The previous callback for garbage collecting client groups has been removed, as it was only triggered when deleting a group with no clients, something we no longer need to handle.

Additionally, the `clientGroupIDs` field in the protocol is now always an empty array. The server already tolerates its absence, and this makes it clearer what the message can contain.