feat(replicache): Expose replicache/expo-sqlite and replicache/op-sqlite (#4887)
This PR reintroduces the `SQLiteStore` abstraction, now using a single connection per store instead of multiple. The previous approach with multiple connections was incompatible with async `op-sqlite`, so this update resolves that issue.

The PR adds and exports two new modules: `replicache/expo-sqlite` and `replicache/op-sqlite`. Each provides an implementation of the `StorageProvider` interface, which can be supplied via the `kvStore` option. This should make it easier to use SQLite-based storage across different environments.

Additionally, there is a third implementation using `@rocicorp/zero-sqlite3`, intended only for testing purposes. It is not exported for production use.

For Zero we export `opSQLiteStoreProvider` and `expoSQLiteStoreProvider` from `@rocicorp/zero/react-native`